% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectMouseMacrophage.R
\docType{class}
\name{TrenaProjectMouseMacrophage-class}
\alias{TrenaProjectMouseMacrophage-class}
\alias{.TrenaProjectMouseMacrophage}
\alias{TrenaProjectMouseMacrophage}
\title{TrenaProjectMouseMacrophage-class}
\usage{
TrenaProjectMouseMacrophage(quiet = TRUE)
}
\value{
An object of the TrenaProjectMouseMacrophage class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
